Implement transactional databases to reduce disk use.
Can be used with io.config.json file. Set autosavemode to commit.

Can also do /ioctrl --sync [dbnames / --all]

This will write out a transaction diff.

+++ io.js
@@ -22,9 +22,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ({
     name: "Standard Input/Output"
     ,
-    authors: ["ArchZombie0x"]
+    require: ["dmp"]
     ,
-    require: []
+    authors: ["ArchZombie0x"]
     ,
     openDBs: null
     ,
@@ -35,14 +35,16 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         this.openDBs = new Object;
         script.registerHandler("step", this);
 
-        this.config = this.readConfig("io", {autosave:60000});
+        this.config = this.readConfig("io", {autosave:60000, autosavemethod: "commit"});
+
+        if (!sys.exists("js_databases")) sys.mkdir("js_databases");
     }
     ,
     read: function (dbname) 
     {
-        if (sys.exists(dbname+".jsqz"))
+        if (sys.exists("js_databases/" + dbname + ".jsqz"))
         {
-            return sys.readObject(dbname + ".jsqz");
+            return sys.readObject("js_databases/" + dbname + ".jsqz");
         }     
 
         return new Object;
@@ -50,7 +52,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     ,
     write: function (dbname, obj, fast)
     {
-        sys.writeObject(dbname + ".jsqz", obj, (fast?1:9));
+        sys.writeObject("js_databases/" +dbname + ".jsqz", obj, (fast?1:9));
     }
     ,
     readConfig: function (cfgname, defaults)
@@ -94,33 +96,93 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         {
             throw new Error("DB already open");//return this.openDBs[dbname].db;
         }
-
-        this.openDBs[dbname] = 
+        var db, patches, dbo, dataText;
+        get_data: 
+        {
+            if (!sys.exists("js_databases/" + dbname + ".jsqz")) 
             {
-                lastSave: +new Date
-                ,
-                db: this.read(dbname)
-                ,
-                hasChanges: null
-                ,
-                referenceCount: null
+                db = new Object;
+                sys.writeObject("js_databases/" +dbname + ".jsqz", db, 9);
+                break get_data;
+            }
+            else
+            {
+                db = sys.readObject("js_databases/" + dbname + ".jsqz");
+            }
+            
+            if (!sys.exists("js_databases/" + dbname + ".jsqz.transactions")) break get_data;
+
+            script.log("Applying transactional patches to transactional database " + dbname);
+
+            var dataText = JSON.stringify(db, null, 1);
+
+            var patches = sys.read("js_databases/" + dbname + ".jsqz.transactions").split(/\n/g);
+
+            if (patches[patches.length - 1] === "") patches.pop();
+
+            var dmp = new this.dmp.diff_match_patch;
+
+            for (var x in patches)
+            {
+                dataText = dmp.patch_apply(JSON.parse(patches[x]), dataText)[0];
             }
 
-        return this.openDBs[dbname].db;
+            db = JSON.parse(dataText);
+
+            this.write( dbname, db);
+
+            sys.rm("js_databases/" + dbname + ".jsqz.transactions");
+        }
+        
+        if (!dataText) dataText = JSON.stringify(db, null, 1);
+
+        dbo = { db: db, lastSave: +new Date, lastCommit: +new Date, dataText: dataText, hasChanges:null };
+
+        this.openDBs[dbname] = dbo;
+
+        return db;
     }
     ,
     flushDB: function (dbname)
     {
         var metadb = this.openDBs[dbname];
         var db = metadb.db;
+        var start = +new Date;
 
         if (metadb.hasChanges === null || metadb.hasChagnes === true)
         {
+            var start = +new Date;
             this.write(dbname, db, true);
-            if (metadb.hasChanges === true) metadb.hasChanges = false;           
+            if (metadb.hasChanges === true) metadb.hasChanges = false; 
+            metadb.dataText = JSON.stringify(db, null, 1);
+            if (sys.exists("js_databases/" + dbname + ".jsqz.transactions")) sys.rm("js_databases/" + dbname + ".jsqz.transactions");
         }
 
         metadb.lastSave = +new Date;
+        var end = +new Date;
+        script.log("Synchronized database " + dbname + ", took " + (end - start) + "ms.");    
+    }
+    ,
+    commitDB: function (dbname)
+    {
+        
+        var start = +new Date;
+        var newData = JSON.stringify(this.openDBs[dbname].db, null, 1);
+
+        var dmp = new this.dmp.diff_match_patch;
+
+        var patch = dmp.patch_make(this.openDBs[dbname].dataText, dmp.diff_lineMode_(this.openDBs[dbname].dataText, newData));
+
+        sys.append("js_databases/" + dbname + ".jsqz.transactions", JSON.stringify(patch) + "\n");
+
+        this.openDBs[dbname].dataText = newData;
+        var end = +new Date;
+        script.log("Commited database " + dbname + ", took " + (end - start) + "ms.");
+    }
+    ,
+    markDB: function (dbname)
+    {
+        this.openDBs[dbname].hasChanges = true;
     }
     ,
     closeDB: function (dbname)
@@ -129,10 +191,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
         var db = metadb.db;
 
         if (!db) return;
-
+        var start = +new Date;
         this.write(dbname, db, false);
 
+        if (sys.exists("js_databases/" +dbname + ".jsqz.transactions")) sys.rm("js_databases/" + dbname + ".jsqz.transactions");
+
         delete this.openDBs[dbname];
+        var end = +new Date;
+        script.log("Closed database " + dbname + ", took " + (end - start) + "ms.");
     }
     ,
     purgeDB: function (dbname)
@@ -150,8 +216,11 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     {
         var metadb = this.openDBs[dbname];
         var db = metadb.db;
-
-        sys.writeObject(dbname + ".backup."+ (+new Date)+".jsqz.bak", db, 9);
+        var start = +new Date;
+        var backupname = dbname + ".backup."+start+".jsqz.bak";
+        sys.writeObject("js_databases/" + backupname, db, 9);
+        var end = +new Date;
+        script.log("Backed up database " + dbname + " to " + backupname + ", took " + (end - start) + "ms.");
     }
     ,
     step: function ()
@@ -162,10 +231,17 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
             if (this.openDBs[x].lastSave + this.config.autosave <= now)
             {
                 var start = +new Date;
-                this.flushDB(x);
+                if (this.config.autosavemethod === "commit")
+                {
+                    this.commitDB(x);
+                }
+                else
+                {
+                    this.flushDB(x);
+                }
                 var end = +new Date;
 
-                script.log("Autosaved database " + x + ", took " + (end-start) + "ms.");
+                script.log((this.config.autosavemethod === "commit" ? "Autocommited" : "Autosaved" ) + " database " + x + ", took " + (end-start) + "ms.");
 
 
                 return;
